requirements.txt
anthropic
python-dotenv
requests
flask
schedule